work in progress

+++ platform/app/utils/messaging/AccountWiper.java
@@ -226,7 +226,7 @@ public class AccountWiper extends AbstractActor {
 			if (user.role == UserRole.PROVIDER) {
 				HPUser hp = HPUser.getById(userId, Sets.create("provider"));
 				user.delete();
-				if (!User.exists(CMaps.map("provider", hp.provider).map("status", User.NON_DELETED))) {
+				if (hp.provider != null && !User.exists(CMaps.map("provider", hp.provider).map("status", User.NON_DELETED))) {
 					OrganizationResourceProvider.deleteOrganization(context, hp.provider);
 					HealthcareProvider.delete(hp.provider);
 				}
